UI FIX: New UI Added as per requirement
